requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
face_recognition==1.3.0
numpy==1.24.3
pillow==9.5.0
python-multipart==0.0.6
aiofiles==23.1.0

langchain==0.0.199
langchain-community==0.1.3
langchain-groq==0.0.1

opencv-python==4.7.0.72
python-dotenv==1.0.0
